python_stuff/requirements.txt
Pillow
OmegaConf
accelerate
pytorch_lightning
numpy
diffusers
transformers
safetensors
invisible-watermark
# gfpgan
# realesrgan
GitPython
pytorch_lightning
torchdiffeq
tdqm
